Chore: moved pages to their respective folders
