+++ frontend/src/app/address-book/layout.tsx
@@ -1,0 +1,18 @@
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Starknet Scaffold",
+  description: "Created with love",
+};
+
+export default function AddressBookRootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <main className="flex min-h-svh flex-col justify-between gap-16">
+      {children}
+    </main>
+  );
+}
