+++ frontend/src/app/layout.tsx
@@ -1,11 +1,8 @@
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
 import "./globals.css";
 import { StarknetProvider } from "~/StarknetProvider";
 import { Toaster } from "react-hot-toast";
 import Footer from "./components/ui_components/Footer";
-
-const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
   title: "Starknet Scaffold",
